Alter calculation to determine tracking from beacon distances
Previously, the distance between beacons was 400 meters, with
accuracy forced to 100 meters max and age to 100 seconds max.

This is now changed to the sum of the accuracies of the two
beacons being compared, PLUS the sum of the ages * 15 m/s,
PLUS 50 meters. This is a much more dynamic calculation that
allows for superior tracker detection using less accurate data
since it reduces the possibility of failing to obtain any
location.