Downgrade Android Gradle plugin version (F-Droid Hotfix)
Hotfix for F-Droid. 

The F-Droid Build Server is not compatible with Android Gradle Plugin Version 8.12. or higher. This is a temporary fix for this release. The Github and Google Play Store Versions of the app are built using the most recent AGP Version. AirGuard Version 2.6.2 is built using AGP Version 8.13 for Github and Google Play.